Create owc-schema.json
This is the first attempt to create JSON validation file. This approach is described here: http://json-schema.org. this is the current version of the draft specification: http://tools.ietf.org/html/draft-zyp-json-schema-04

I'm using ValidatorBuddy to test some of the simple examples that Pedro provided. 

I think the results are promising and could be part of the OWC JSON specification.